Reverted to two separate code coverage test files
